readme
User Manual

1. Aim of the game is for each player to bring his or her checkers into
his or her home board
and then to bear them off the board. First player that has bearded all
his or her checkers off
the board wins the game.
2. The gameboard consists of 24 triangular-shaped fields called points.
Points are numbered
from 1 to 24 starting from the bottom right corner of the gameboard. The
triangles alternate
in color over one and are grouped into four quadrants of six fields in
each. Each player has
15 checkers of the same color placed on various fields. Note:
Gameboard’s initial layout is
sketched in the „Gameboard layout“ section of this document.
3. The movement of the checkers is determined by the number of points
obtained in rolling
of dice. The checkers movement is always forward. A player can move his
or her checkers
only to an open point. A point containing at least two opponent’s
checkers is considered safe
and not opened. The player can either move two checkers with two
separate moves or one
checker with the sum of both dice.
4. A point occupied by a single checker is called a blot. If an
oppposing checker lands on a
blot then the point is hit and checker which was on it gets placed on
the bar. If a player
has at least on checker on the bar, his or her obligation is to enter
those checkers into the
opponent’s home board. A checker has to be moved to an open point, a
point that is not
occupied by at least two or more opponent’s checkers, corresponding to
one of the rolled
dice. If neither of the points is opened the player has to forfeit his
or her turn. If a player is
able to enter some of his or her checkers, he or she must enter as many
checkers as possible
and forfeit the remaining turns. After all of the checkers on the bar
have been entered any
unused number of steps must be played by the rules defined.
5. Once a player has moved all of his or her checkers to his or her home
board he or she can
start bearing them off. A checker is bearded off by rolling a number
that corresponds to the
point on which the checker resides and is removed from the board. If
there are no checkers
on the point indicated by the roll the player must make a legal move
using a checker on a
higher-numbered point. If there are no checkers on the higher numbered
points the player
can remove a checker from the next highest point but is not obligated to
do that if he or she
can otherwise make a legal move.